Refactoring TDX specific types to correct places
- `verify` function now is part of `ReportMac` instead of
  `TdxReportV1`, although maintaining compatibility that the `verify`
  function still exists in `TdxReportV1`
- Move the internal `tdx_arch` module to be part of entire `arch`
  module in the crate.
- Move `ReportMac` to the main crate level instead of under `tdx`
  module namespace
- Move all `TdxError` crates out as it is more relevant to `tdx-ql`
  crate. The error from SGX-ISA should be only the `ErrorCode` types.
  `TdxError` is not part of the ISA.
